requirements.txt
python-dotenv 
folium
streamlit
streamlit_folium
langchain
langchain_core
langchain_community
langchain_groq
langchain_huggingface
langgraph
faiss-cpu
nltk
unstructured[all-docs]
pillow
lxml
pydantic
poppler-utils